Use the same Title component for index and show
