+++ app/javascript/blacklight/layouts/showPage.vue
@@ -1,33 +1,32 @@
 <template>
   <section class="container">
-    <h1 v-html="result.data.attributes.title_tesi.attributes.value"></h1>
-    <dl>
-    <template v-for="(property, index) in result.data.attributes">
-      <dt v-html="property.attributes.label"></dt>
-      <dd v-html="property.attributes.value"></dd>
-    </template>
-    </dl>
+    <h1><Title v-bind:item="item"></Title></h1>
+    <PropertyList v-bind:item="item"></PropertyList>
   </section>
 </template>
 
 
 <script>
+import Title from 'blacklight/result/title.vue'
+import PropertyList from 'blacklight/result/property_list.vue'
 
 export default {
   components: {
+    Title,
+    PropertyList
   },
   data() {
     return {
-      result: { data: { attributes: {} } }
+      item: { attributes: {} }
     }
   },
   created() {
     const endpoint = `/catalog/${encodeURIComponent(this.$route.params.id)}`
     this.$http.get(endpoint).then(function(response){
-        this.result = response.data
-        this.message = null
+        this.item = response.data.data
     }, function(error){
         console.error(error.statusText);
+        alert("There was an error retrieveing this record")
     })
   }
 }
