+++ app/javascript/blacklight/result/property_list.vue
@@ -1,0 +1,17 @@
+<template>
+  <dl>
+    <template v-for="(property, index) in item.attributes">
+      <dt v-html="property.attributes.label"></dt>
+      <dd v-html="property.attributes.value"></dd>
+    </template>
+  </dl>
+</template>
+
+<script>
+export default {
+  props: ['item']
+}
+</script>
+
+<style scoped>
+</style>
